requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pymongo==4.6.0
motor==3.3.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
qrcode[pil]==7.4.2
pillow==10.1.0
twilio==8.10.0
sendgrid==6.10.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-decouple==3.8
cryptography==41.0.8 